+++ src/app/login/login-form/types.ts
@@ -1,0 +1,5 @@
+export type signInFieldsType = {
+  email: string;
+  password: string;
+  isPersistent: boolean;
+};
